Resolvers file structure fixes
