+++ app/graphql/resolvers/cars/car_resolver.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Resolvers
+  module Cars
+    class CarResolver < Resolvers::BaseResolver
+      description 'car by id'
+
+      type Types::CarType, null: false
+
+      argument :id, ID, required: true
+
+      def resolve(id:)
+        Car.find(id)
+      end
+    end
+  end
+end
