+++ app/graphql/resolvers/cars_owners_documents/cars_owners_document_resolver.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Resolvers
+  module CarsOwnersDocuments
+    class CarsOwnersDocumentResolver < Resolvers::BaseResolver
+      description 'document'
+
+      type Types::CarsOwnersDocumentType, null: false
+
+      argument :id, ID, required: true
+
+      def resolve(id:)
+        CarsOwnersDocument.find(id)
+      end
+    end
+  end
+end
