--- app/graphql/resolvers/cars_resolver.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Resolvers
-  class CarsResolver < Resolvers::BaseResolver
-    description 'all cars'
-
-    type [Types::CarType], null: false
-
-    def resolve
-      Car.all
-    end
-  end
-end
