+++ app/graphql/resolvers/users/user_resolver.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Resolvers
+  module Users
+    class UserResolver < Resolvers::BaseResolver
+      description 'user by id'
+
+      type Types::UserType, null: false
+
+      argument :id, ID, required: true
+
+      def resolve(id:)
+        User.find(id)
+      end
+    end
+  end
+end
