--- app/graphql/resolvers/users_resolver.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Resolvers
-  class UsersResolver < Resolvers::BaseResolver
-    description 'all users'
-
-    type [Types::UserType], null: false
-
-    def resolve
-      User.all
-    end
-  end
-end
